backend import af

+++ ChronoZoom/ChronoZoom/ChronoZoom.Backend/Business/Interfaces/IBatchService.cs
@@ -8,6 +8,6 @@ namespace ChronoZoom.Backend.Business.Interfaces
 {
     public interface IBatchService
     {
-        int ProcessFile(string filename);
+        long ProcessFile(string filename);
     }
 }
